+++ app/tests/test_layout.py
@@ -4,11 +4,6 @@ Tests for the timeline layout functionality.
 
 from datetime import datetime
 
-from app.extract_timeline_spans import (
-    _get_date_range,
-    generate_week_axis,
-    _extract_timeline_spans,
-)
 from app.utils import compute_week_index
 
 
@@ -36,78 +31,3 @@ def test_compute_week_index():
     assert (
         compute_week_index(datetime.strptime("2023-12-31", "%Y-%m-%d"), min_date) == 52
     )
-
-
-def test_get_date_range(sample_entries):
-    """Test getting min and max dates from entries."""
-    min_date, max_date = _get_date_range(sample_entries)
-
-    # Min date should be adjusted to start of week
-    assert min_date.year == 2021
-    assert min_date.month == 2
-    assert min_date.day == 1
-
-    # Max date should be adjusted to end of week
-    assert max_date.year == 2021
-    assert max_date.month == 10
-    assert max_date.day == 4
-
-
-def test_generate_week_axis():
-    """Test generating week axis DataFrame."""
-    min_date = datetime(2023, 1, 1)
-    max_date = datetime(2023, 1, 31)
-
-    weeks_df = generate_week_axis(min_date, max_date)
-
-    # Should have 5 weeks (Jan 1-7, 8-14, 15-21, 22-28, 29-31)
-    assert len(weeks_df) == 5
-    assert weeks_df["week_index"].tolist() == [0, 1, 2, 3, 4]
-    assert all(weeks_df["year"] == 2023)
-
-    # Check first and last week
-    assert weeks_df.iloc[0]["start_date"] == min_date
-    assert weeks_df.iloc[-1]["end_date"] >= max_date
-
-
-def test_extract_timeline_spans(sample_entries):
-    """Test extracting timeline spans from entries."""
-    spans, min_date, max_date = _extract_timeline_spans(sample_entries)
-
-    # Check min and max dates
-    assert min_date is not None
-    assert max_date is not None
-    assert min_date < max_date
-
-    # Check spans
-    assert len(spans) > 0
-    assert "entry_idx" in spans[0]
-
-
-def test_extract_timeline_spans_empty():
-    """Test extracting timeline spans with empty entries."""
-    spans, min_date, max_date = _extract_timeline_spans([])
-
-    # Check min and max dates
-    assert min_date is None
-    assert max_date is None
-    assert len(spans) == 0
-
-
-def test_extract_timeline_spans_missing_dates():
-    """Test preparing timeline data with entries missing dates."""
-    entries = [
-        {
-            "title": "No Dates Entry",
-            "canonical_title": "No Dates Entry",
-            "type": "Movie",
-            "tags": {},
-        }
-    ]
-
-    spans, min_date, max_date = _extract_timeline_spans(entries)
-
-    # Should have weeks but no spans
-    assert min_date is None
-    assert max_date is None
-    assert len(spans) == 0
